SegmentOptions needs to be pointer in SegmentRequest struct
MailChimp will only accept `static_segment` (note: corrected from
`static_segments`) or `options` but not both. Using a struct instead of
a pointer for options will cause it to always be present, which causes
MC to return a 400 if static_segment is also present. Setting it to a
pointer allows it to be nil, which means it can be left empty if we're
doing a static segment